backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.5.3
python-multipart==0.0.9

# Database and authentication
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
python-dotenv==1.0.0

# AI and ML - optimized versions
langchain==0.2.17
langchain-community==0.2.19
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==4.1.0
transformers==4.41.2
huggingface_hub==0.23.0

# Document processing
pypdf==5.6.0

# HTTP requests
requests==2.31.0

# Core scientific computing
numpy==1.24.3

# Memory monitoring
psutil==5.9.8
